requirements.txt
streamlit==1.8.0
wikipedia-api==0.5.4
